+++ examples/data-sources/meraki_insight_monitored_media_server/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_insight_monitored_media_server" "example" {
+  organization_id = "123456"
+}