+++ examples/resources/meraki_insight_monitored_media_server/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_insight_monitored_media_server.example "<organization_id>"