+++ docs/guides/changelog.md
@@ -18,6 +18,7 @@ description: |-
 - Add `meraki_camera_video_settings` resource and data source
 - Add `meraki_camera_wireless_profile` resource and data sources
 - Add `meraki_camera_device_wireless_profiles` resource and data source
+- Add `meraki_insight_monitored_media_server` resource and data source
 
 ## 0.1.6